security/manager/ssl/tests/mochitest/bugs/test_ev_validation.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=813418
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 813418</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=813418">Mozilla Bug 813418</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  var wnd = window.open("https://ev-valid.example.com/tests/security/manager/ssl/tests/mochitest/bugs/test_ev_validation_child.html");
  window.addEventListener("message", function(event) {
     if (event.origin == "https://ev-valid.example.com") {
       is(event.data, "EV", "Child was EV valid.");
       wnd.close();
       wnd = window.open("https://example.com/tests/security/manager/ssl/tests/mochitest/bugs/test_ev_validation_child.html");
     }
     else if (event.origin == "https://example.com") {
       is(event.data, "secure", "Child was just secure (NO EV) no ev flags present.");
       wnd.close();
       wnd = window.open("https://ev-invalid.example.com/tests/security/manager/ssl/tests/mochitest/bugs/test_ev_validation_child.html");
     }
     else if (event.origin == "https://ev-invalid.example.com") {
       is(event.data, "secure", "Child was just secure (NO EV ev cert from invalid anchor).");
       wnd.close();
       SimpleTest.finish(); //expected end of tests.
     }
     else{
       ok(false,"something is broken");
       SimpleTest.finish();
     }

    }, false);

</script>
</pre>
</body>
</html>
